+++ tests/testthat/test-standard-error.R
@@ -95,7 +95,9 @@ test_that("Check that variance estimation are coherent in PLNfit",  {
       trace           = 2
     )
 
-  myPLN$postTreatment(Y, X, exp(log_O), config = config_post)
+  config_optim <- config_default_nlopt
+  myPLN$postTreatment(Y, X, log_O, config_post = config_post, config_optim = config_optim)
+
 
   tr_variational <- sum(standard_error(myPLN, "variational")^2)
   tr_bootstrap   <- sum(standard_error(myPLN, "bootstrap")^2)